Добавил контролы, и гифку в окне авторизации (получилось гавной, если ты как-то решишь эту проблему будет хорошо)

+++ adonetCourseProject/adonetCourseProject/Warehouse.cs
@@ -9,7 +9,7 @@ namespace adonetCourseProject
     public class Warehouse
     {
         public int Id { get; set; }
-        public decimal Quantity { get; set; }
+        public int Quantity { get; set; }
         public virtual Product Product { get; set; }
     }
 }
